requirements-dev.txt
pytest==2.6.3
pytest-pythonpath==0.3
tox==1.8.0
pytest-httpbin==0.0.3
requests==2.4.3
factory-boy==2.4.1
